backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
bcrypt
python-jose[cryptography]
python-multipart
psycopg2-binary
python-dotenv
inngest
# anthropic
google-generativeai
pypdf2
python-docx
